refactor(auto_source): split ClassClustering overlap and scoring
Orchestrator collects class groups; OverlapResolver owns layout-wrapper
and 1-to-1 nesting filters; GroupScorer owns heading/date/word ranking.
Public .call API and Sets for discarded/tag membership stay unchanged.
